+++ React InstantSearch/injected-content/package.json
@@ -13,6 +13,8 @@
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "react-instantsearch-dom": "6.18.0",
+    "react-router": "6.2.1",
+    "react-router-dom": "6.2.1",
     "react-scripts": "2.1.1"
   },
   "devDependencies": {
