+++ React InstantSearch/injected-content/src/routes.js
@@ -1,0 +1,9 @@
+import { InterspersedHits } from './pages/InterspersedHits';
+
+export const routes = [
+  {
+    path: '/interspersed-hits',
+    label: 'Interspersed hits',
+    element: InterspersedHits,
+  },
+];
