+++ React InstantSearch/injected-content/src/searchClient.js
@@ -1,0 +1,6 @@
+import algoliasearch from 'algoliasearch/lite';
+
+export const searchClient = algoliasearch(
+  'latency',
+  'b51f8ae5601632c6e32a0f8c6f3f6971'
+);
